.claude/worktrees/clever-tesla/requirements.txt
# Roamin Ambient Agent — Python dependencies
# Install order matters for llama-cpp-python (see NOTE below)

# --- Core inference ---
# NOTE: llama-cpp-python requires manual CUDA build on this machine.
# Run scripts/install_llama_cpp_cuda.ps1 after pip install -r requirements.txt
# llama-cpp-python  <- installed separately

# --- Memory ---
chromadb>=0.5.0

# --- Voice ---
pyttsx3>=2.90
sounddevice>=0.4.6
openai-whisper>=20231117
keyboard>=0.13.5

# --- Screen observation ---
Pillow>=10.0.0

# --- Voice Activity Detection ---
silero-vad>=5.1.2

# --- HTTP (model fallback) ---
requests>=2.31.0

# --- Web search ---
duckduckgo-search>=8.0.0

# --- GPU / Vision ---
torch>=2.6.0
numpy>=1.26.0

# --- Windows ---
pywin32>=306

# --- Dev / Test ---
pytest>=8.0.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
